+++ packages/component-boilerplate/src/index.ts
@@ -30,7 +30,10 @@ export {
   GRAPHQL_CONTEXT,
 } from "./hooks/graphql/useGraphQLClient";
 export { useRefreshDispatcher } from "./hooks/graphql/Refresh";
-export { useParentQuery } from "./hooks/graphql/useParentQuery";
+export {
+  useParentQuery,
+  useParentQueryValue,
+} from "./hooks/graphql/useParentQuery";
 
 //      Navigation
 export { navigation, useCurrentPage } from "./hooks/useNavigation";
@@ -50,7 +53,12 @@ export {
   useHost,
   setImplementation as setUseHostImplementation,
 } from "./hooks/useHost";
-export { useParentState } from "./hooks/useParentState";
+export {
+  useParentState,
+  useParent,
+  useParentValue,
+  useSetParent,
+} from "./hooks/useParentState";
 
 //
 //      Common higher-level functionality